Refactor fetch into App
