+++ src/pages/Dashboard/index.jsx
@@ -1,23 +1,10 @@
-import { useEffect, useState } from 'react'
 import PeopleList from './components/PeopleList'
 
 function Dashboard(props) {
-  const { hiredPeople } = props
-
-  const [people, setPeople] = useState([])
-
-  useEffect(() => {
-    fetch('https://randomuser.me/api/?results=5')
-      .then(response => response.json())
-      .then(data => {
-        const dataWithId = data.results.map((p, id) => {
-          p.id = id
-          return p
-        })
-        console.log(dataWithId)
-        setPeople(dataWithId)})
-  }, [])
-
+  // const [people, setPeople] = useState([])
+  const people = props.people
+  const hiredPeople = props.hiredPeople
+  
   return (
     <main className="dashboard-layout">
       <section>
